+++ AdventOfCode2019/Opcodes/WriteOutput.cs
@@ -10,7 +10,7 @@ namespace AdventOfCode2019.Opcodes
 
         protected override void InnerExecute(IntcodeContext context)
         {
-            int output = context.Read(1);
+            long output = context.Read(1);
             if (!context.Mode.HasFlag(IntcodeMode.Quiet))
                 Console.WriteLine("Output: " + output);
             context.OutputQueue.Enqueue(output);
